+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.23.0
+* Added supported for creating and deleting breakpoints from the REPL. Use `new 123` to create a breakpoint in line 123, and `del 123` to delete it.
+* Use 1.24.0 version of Debug Adapter Protocol and libraries
+
 ## 0.22.0
 * Refactored the 'Mock Debugger' functionality into a separate class. This makes it more obvious how a debug adapter 'adapts' to a debugger or runtime.